helpfuls/math/sigma.pi/requirements.txt
numexpr
numpy
cupy-cuda12x